+++ src/sagas/index.js
@@ -1,2 +1,3 @@
 export { withAuthSession, getAuthStateChannel } from './authState';
 export { default as configure } from './configure';
+export { default as createExpirationDate } from './utilities/createExpirationDate';